# Eliminate allocations from ReflectionUtility.GetImplementedInterfaces