--- a/YamlDotNet.Benchmark/SerializationBenchmarks.cs
+++ b/YamlDotNet.Benchmark/SerializationBenchmarks.cs
@@ -19,7 +19,36 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using BenchmarkDotNet.Running;
-using YamlDotNet.Benchmark;
+using System.Text;
+using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
+using YamlDotNet.Serialization;
 
-BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+namespace YamlDotNet.Benchmark;
+
+[MemoryDiagnoser]
+[MediumRunJob(RuntimeMoniker.Net80)]
+[MediumRunJob(RuntimeMoniker.Net47)]
+public class SerializationBenchmarks
+{
+    public class SampleRecord
+    {
+        public SampleRecord(string name, string description)
+        {
+            Name = name;
+            Description = description;
+        }
+
+        public string Name { get; private set; }
+        public string Description { get; private set; }
+    }
+
+    private readonly IReadOnlyCollection<SampleRecord> configs = Enumerable.Range(0, 10_000).Select(i => new SampleRecord("MyName", "MyDescription")).ToList();
+    private readonly ISerializer serializer = new SerializerBuilder().DisableAliases().Build();
+
+    [Benchmark]
+    public string Serializer()
+    {
+        return serializer.Serialize(configs);
+    }
+}